The TransformAction doesn't do concurrency
The semaphore was removed; the task of running multiple transform
actions is left on the resource pool.

There was still an error with the actPool release; which the MainIT and
the asserts (in the TransformAction!, run with `-ea`) helped to discover.

Main invokes the executor in a different manner; the old way didn't wait
for workers to finish (at least when run through IDE).

The Configuration and ActionSequence have been changed in order to pass
@max-jobs as resource pool size for the transform action