Added failing test for worker
In my opinion it should not proceed with another action sequence when
there were errors.
If there are network/server issues the next sequence will either fail
too; or you'll get the metadata in a mix of formats (as all should
support oai_dc and it can be expected that oai_dc will have the lowest
priority).
It's questionable with ListIdentifiers scenario - one malformed record
can stop the whole harvest. But why is it malformed? If it is bad xml,
when fixed, it should update the last modified time (to support
incremental); if it's only due to a temporary issue; then it might not
be possible for incremental harvest to get the resource next time.