New test
verifies that all 5 records provided in dspace5l ListIdentifiers response are saved.
There was actually an issue with the test data, where the identifier in
the GetRecord response for TOKENIZER+TAGGER was missing the +TAGGER
part. Without the +TAGGER part it matches another id so two records were
written into the same file