Keep deleted ids in Provider, pass them from StripAction
StripAction:
There should never be status="deleted" & metadata element. Status can be
either "deleted" or the attribute is not present at all
so this check:
```
if (status == null || !status.equals("deleted"))
```
in the part of code where the parser is reading the metadata element is
superfluous. Might bring it back if we want to check the spec is
followed.